Rename package to foxglove
