+++ examples/downloading_data.py
@@ -1,7 +1,7 @@
 import datetime
 from datetime import datetime, timedelta
 
-from foxglove_data_platform.client import Client
+from foxglove.client import Client
 
 token = "<YOUR API TOKEN HERE>"
 device_id = "<YOUR DEVICE ID>"
