+++ examples/events.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from foxglove_data_platform.client import Client
+from foxglove.client import Client
 
 token = "<YOUR API TOKEN>"
 device_id = "<YOUR DEVICE ID>"
