+++ setup.cfg
@@ -1,10 +1,10 @@
 [metadata]
-name = foxglove-data-platform
-version = 0.14.0
-description = Client library for Foxglove Data Platform.
+name = foxglove
+version = 1.0.0
+description = Client library for the Foxglove API.
 long_description = file: README.md
 long_description_content_type = text/markdown
-url = https://github.com/foxglove/py-data-platform
+url = https://github.com/foxglove/foxglove-python
 classifiers =
     Programming Language :: Python :: 3
     License :: OSI Approved :: MIT License
@@ -21,4 +21,4 @@ packages = find:
 python_requires = >=3.7
 
 [options.package_data]
-foxglove_data_platform = py.typed
+foxglove = py.typed
