+++ tests/test_coverage.py
@@ -2,7 +2,7 @@ from datetime import datetime
 
 import responses
 from faker import Faker
-from foxglove_data_platform.client import Client
+from foxglove.client import Client
 
 from .api_url import api_url
